# api/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: BFF API
    version: 0.0.1
paths:
    /demo/v1/trainer/getTrainerAvailableHours:
        post:
            tags:
                - BFF
            operationId: BFF_GetTrainerAvailableHours
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/bff.GetTrainerAvailableHoursRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/bff.GetTrainerAvailableHoursRespone'
    /demo/v1/trainer/makeHourAvailable:
        post:
            tags:
                - BFF
            operationId: BFF_MakeHourAvailable
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/bff.MakeHourAvailableRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /demo/v1/trainer/makeHourUnavailable:
        post:
            tags:
                - BFF
            operationId: BFF_MakeHourUnavailable
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/bff.MakeHourUnavailableRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /demo/v1/training/approveRescheduleTraining:
        post:
            tags:
                - BFF
            operationId: BFF_ApproveRescheduleTraining
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/training.ApproveRescheduleTrainingRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /demo/v1/training/cancelTraining:
        post:
            tags:
                - BFF
            operationId: BFF_CancelTraining
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/training.CancelTrainingRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /demo/v1/training/createTraining:
        post:
            tags:
                - BFF
            operationId: BFF_CreateTraining
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/training.CreateTrainingRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/training.CreateTrainingResponse'
    /demo/v1/training/getTraining:
        post:
            tags:
                - BFF
            operationId: BFF_GetTraining
            requestBody:
                content:
                    application/json: {}
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/training.GetTrainingResponse'
    /demo/v1/training/rejectRescheduleTraining:
        post:
            tags:
                - BFF
            operationId: BFF_RejectRescheduleTraining
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/training.RejectRescheduleTrainingRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /demo/v1/training/requestRescheduleTraining:
        post:
            tags:
                - BFF
            operationId: BFF_RequestRescheduleTraining
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/training.RequestRescheduleTrainingRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /demo/v1/training/rescheduleTraining:
        post:
            tags:
                - BFF
            operationId: BFF_RescheduleTraining
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/training.RescheduleTrainingRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
components:
    schemas:
        bff.GetTrainerAvailableHoursRequest:
            type: object
            properties:
                dateFrom:
                    type: string
                    format: date-time
                dateTo:
                    type: string
                    format: date-time
        bff.GetTrainerAvailableHoursRespone:
            type: object
            properties:
                dates:
                    type: array
                    items:
                        $ref: '#/components/schemas/bff.GetTrainerAvailableHoursRespone_Date'
        bff.GetTrainerAvailableHoursRespone_Date:
            type: object
            properties:
                date:
                    type: string
                    format: date-time
                HasFreeHours:
                    type: boolean
                hours:
                    type: array
                    items:
                        $ref: '#/components/schemas/bff.GetTrainerAvailableHoursRespone_Hour'
        bff.GetTrainerAvailableHoursRespone_Hour:
            type: object
            properties:
                available:
                    type: boolean
                hasTrainingScheduled:
                    type: boolean
                hour:
                    type: string
                    format: date-time
        bff.MakeHourAvailableRequest:
            type: object
            properties:
                time:
                    type: array
                    items:
                        type: string
                        format: date-time
        bff.MakeHourUnavailableRequest:
            type: object
            properties:
                time:
                    type: array
                    items:
                        type: string
                        format: date-time
        training.ApproveRescheduleTrainingRequest:
            type: object
            properties:
                trainingUuid:
                    type: string
        training.CancelTrainingRequest:
            type: object
            properties:
                trainingUuid:
                    type: string
        training.CreateTrainingRequest:
            type: object
            properties:
                notes:
                    type: string
                time:
                    type: string
                    format: date-time
        training.CreateTrainingResponse:
            type: object
            properties:
                trainingUuid:
                    type: string
        training.GetTrainingResponse:
            type: object
            properties:
                training:
                    type: array
                    items:
                        $ref: '#/components/schemas/training.GetTrainingResponse_Training'
        training.GetTrainingResponse_Training:
            type: object
            properties:
                canBeCancelled:
                    type: boolean
                moveProposedBy:
                    type: string
                moveRequiresAccept:
                    type: boolean
                notes:
                    type: string
                proposedTime:
                    type: string
                    format: date-time
                time:
                    type: string
                    format: date-time
                user:
                    type: string
                userUuid:
                    type: string
                uuid:
                    type: string
        training.RejectRescheduleTrainingRequest:
            type: object
            properties:
                trainingUuid:
                    type: string
        training.RequestRescheduleTrainingRequest:
            type: object
            properties:
                trainingUuid:
                    type: string
                notes:
                    type: string
                time:
                    type: string
                    format: date-time
        training.RescheduleTrainingRequest:
            type: object
            properties:
                trainingUuid:
                    type: string
                notes:
                    type: string
                time:
                    type: string
                    format: date-time
tags:
    - name: BFF